ingestion/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.1
minio==7.2.5
numpy==1.26.3
opencv-python==4.9.0.80
python-dotenv==1.0.1 